refactor: rename query-expansion package to query
Shorter, cleaner package name: `query` instead of `query-expansion`.
Python namespace: `rag_facile.query` (was `rag_facile.query_expansion`).

- Renamed packages/query-expansion/ → packages/query/
- Renamed src/rag_facile/query_expansion/ → src/rag_facile/query/
- Updated all internal imports across the package
- Updated all external references:
  - packages/pipelines/pyproject.toml
  - packages/rag-facile-lib/pyproject.toml (force-include path)
  - pyproject.toml (root workspace)
  - release-please-config.json
  - AlbertPipeline import
  - retrieval package docstring
