+++ packages/query/README.md
