+++ packages/query/src/rag_facile/query/multi_query.py
@@ -16,8 +16,8 @@ from __future__ import annotations
 import logging
 from typing import TYPE_CHECKING, Any
 
-from rag_facile.query_expansion._base import QueryExpander
-from rag_facile.query_expansion._models import ExpandedQueries
+from rag_facile.query._base import QueryExpander
+from rag_facile.query._models import ExpandedQueries
 
 
 if TYPE_CHECKING:
